Bring check_relativistic back to _boilerPlate
As it turns out, when we added z_mean to the arguments to coulomb_log,
we changed their order. z_mean now came in place of the old V argument.
And that means all the time, we were going with the default nan
velocity. Thus, the one for _check_relativistic was being calculated
off of the temperature (low for these cases). Thus, the warning
was correctly not firing - the velocity was too low for it to launch!